refactor: move Breadcrumbs to a separate folder
